We now have three sloth engines: Torpor, Apathy, Inertia.
I discovered that the original Apathy circuit has some
undesirable properties for simulation. The main problem
is that it pushes op-amps that my code expects to be
in linear mode all the way to the rails. This messes up
my simulation theory. Andrew Fitch confirmed this is
the behavior of the real circuit.

I have decided to take a different route. I am creating
Apathy and Inertia variants by introducing a time dilation
factor. Thus Apathy and Inertia act like Torpor, only
at progressively slower speeds.

To prevent them from having identical initial trajectories,
I also add a slight variation to the initial charge of
the C3 capacitor, which is represented by the voltage "w".
The 3 variants pass in a different value w0 to initialize
the "w" voltage slightly differently. This causes the initial
trajectory slope to be slightly different. The Butterfly Effect
takes care of the future trejectory unfolding differently.