Make the indicator light vary in brightness depending on voltage.
Instead of instantly toggling the indicator light, make it get
brighter or dimmer depending on the amplitude of the y voltage.
This is closer to the spirit of the Triple Sloth implementation,
which I will start soon. I want Triple Sloth to be a straightforward
combination of the simple sloths.